requirements.txt
mysql-connector-python
streamlit
google-generativeai
python-dotenv